+++ package.json
@@ -12,9 +12,11 @@
     "@types/react-dom": "^18.2.4",
     "axios": "^1.4.0",
     "dotenv": "^16.0.3",
+    "http-proxy-middleware": "^2.0.6",
     "react": "^18.2.0",
     "react-daum-postcode": "^3.1.1",
     "react-dom": "^18.2.0",
+    "react-error-overlay": "^6.0.9",
     "react-router-dom": "^6.11.1",
     "react-scripts": "5.0.1",
     "sass": "^1.62.1",
