+++ src/components/Weather/Clock.tsx
@@ -1,0 +1,17 @@
+import React, { useState, useEffect } from 'react';
+
+export function Clock() {
+  const [currentTime, setCurrentTime] = useState(new Date());
+
+  useEffect(() => {
+    const timer = setInterval(() => {
+      setCurrentTime(new Date());
+    }, 1000);
+
+    return () => {
+      clearInterval(timer);
+    };
+  }, []);
+
+  return <>{currentTime.toLocaleTimeString()}</>;
+}
